susc working with momentum, still not parity

+++ src/helperFunctions.cpp
@@ -145,7 +145,7 @@ void printEnergies(const vector<double> & v) {
     }
 }
 
-int fact(int n) {
+long fact(int n) {
     if ((n==0)||(n==1))
         return 1;
     else
